fix: move get_current_user and get_db to auth.security module
- Moved get_current_user function from main.py to auth/security.py
- Moved get_db function to auth/security.py
- Added necessary imports (OAuth2PasswordBearer, TokenData) to auth/security.py
- Updated main.py to import get_current_user and get_db from auth.security
- Removed duplicate function definitions from main.py
- Fixed import errors in invite_codes router
