+++ backend/auth/security.py
@@ -3,6 +3,9 @@ from typing import Optional
 from jose import JWTError, jwt
 from passlib.context import CryptContext
 from pydantic import BaseModel
+from fastapi import HTTPException, Depends, status
+from fastapi.security import OAuth2PasswordBearer
+from sqlalchemy.orm import Session
 import os
 from dotenv import load_dotenv
 import uuid
@@ -19,6 +22,9 @@ REFRESH_TOKEN_EXPIRE_DAYS = 7
 MAX_LOGIN_ATTEMPTS = 5
 LOCKOUT_DURATION_MINUTES = 15
 
+# OAuth2スキーマ
+oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")
+
 # Redis接続
 redis_client = redis.Redis(
     host=os.getenv("REDIS_HOST", "localhost"),
@@ -33,6 +39,10 @@ class Token(BaseModel):
     access_token: str
     refresh_token: str
     token_type: str
+
+
+class TokenData(BaseModel):
+    username: Optional[str] = None
 
 def verify_password(plain_password: str, hashed_password: str) -> bool:
     return pwd_context.verify(plain_password, hashed_password)
@@ -143,3 +153,36 @@ def log_security_event(event_type: str, username: str, details: str, ip_address:
     redis_client.lpush("security_logs", str(event))
     # 最大1万件までログを保持
     redis_client.ltrim("security_logs", 0, 9999)
+
+
+def get_db():
+    """データベースセッションを取得"""
+    from database import SessionLocal
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
+
+
+def get_current_user(token: str = Depends(oauth2_scheme), db: Session = Depends(get_db)):
+    """現在のユーザーを取得"""
+    from models.user import User
+    
+    credentials_exception = HTTPException(
+        status_code=status.HTTP_401_UNAUTHORIZED,
+        detail="無効な認証情報",
+        headers={"WWW-Authenticate": "Bearer"},
+    )
+    try:
+        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
+        username: str = payload.get("sub")
+        if username is None:
+            raise credentials_exception
+        token_data = TokenData(username=username)
+    except JWTError:
+        raise credentials_exception
+    user = db.query(User).filter(User.username == token_data.username).first()
+    if user is None:
+        raise credentials_exception
+    return user
